Login_Register/requirements.txt
mysql-connector-python
flask
dotenv